Update: split into lexer and internal directory
