+++ lexer/lexer.c
@@ -1,86 +1,5 @@
 #include "lexer.h"
 
-// inputに対して新しくlexer構造体を作成する
-void	new_lexer(t_lexer **l, char *input)
-{
-	*l = x_malloc(sizeof(**l));
-	(*l)->input = input;
-	(*l)->position = 0;
-	(*l)->read_position = 0;
-	(*l)->is_subshell = false;
-	(*l)->is_redirect = false;
-}
-
-// token解析のための分岐処理
-t_token *next_token(t_lexer *l)
-{
-	t_token	*token;
-
-	token = skip_space(l);
-	if (token)
-		return (token);
-	if (l->ch == '|')
-	{
-		if (l->input[l->read_position] == '|')
-		{
-			token = new_token(OR_IF, l, 2, l->position);
-			read_char(l);
-		}
-		else
-			token = new_token(PIPE, l, 1, l->position);
-	}
-	else if (l->ch == '>')
-	{
-		if (l->input[l->read_position] == '>')
-		{
-			token = new_token(REDIRECT_APPEND, l, 2, l->position);
-			read_char(l);
-		}
-		else
-			token = new_token(REDIRECT_OUT, l, 1, l->position);
-	}
-	else if (l->ch == '<')
-	{
-		if (l->input[l->read_position] == '<')
-		{
-			token = new_token(HEREDOC, l, 2, l->position);
-			read_char(l);
-		}
-		else
-			token = new_token(REDIRECT_IN, l, 1, l->position);
-	}
-	else if (l->ch == '&')
-	{
-		if (l->input[l->read_position] == '&')
-		{
-			token = new_token(AND_IF, l, 2, l->position);
-			read_char(l);
-		}
-		else
-			token = new_token(ILLEGAL, l, 1, l->position);
-	}
-	else if (l->ch == '(')
-	{
-		token = new_token(LPAREN, l, 1, l->position);
-		l->is_subshell = true;
-	}
-	else if (l->ch == ')')
-	{
-		token = new_token(RPAREN, l, 1, l->position);
-		l->is_subshell = false;
-	}
-	else if (l->ch == '\0')
-		token = new_token(EOL, l, 1, l->position);
-	else
-	{
-		token = new_token_string(l);
-		return (token);
-	}
-	read_char(l);
-	return (token);
-}
-
-// token_listの先頭アドレスを返す
 t_token	*lex(char *input)
 {
 	t_token	token;
